Fix codeclimate errors
