+++ src/main/java/ru/mail/polis/lsm/artem_drozdov/LsmDAO.java
@@ -8,8 +8,6 @@ import javax.annotation.Nullable;
 import javax.annotation.concurrent.GuardedBy;
 import java.io.IOException;
 import java.io.UncheckedIOException;
-import java.lang.ref.Cleaner;
-import java.lang.ref.PhantomReference;
 import java.nio.ByteBuffer;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -18,7 +16,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.NavigableMap;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.SortedMap;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.ConcurrentSkipListMap;
@@ -129,7 +126,7 @@ public class LsmDAO implements DAO {
     }
 
     private static Iterator<Record> merge(List<Iterator<Record>> iterators) {
-        if (iterators.size() == 0) {
+        if (iterators.isEmpty()) {
             return Collections.emptyIterator();
         }
         if (iterators.size() == 1) {
@@ -144,44 +141,7 @@ public class LsmDAO implements DAO {
     }
 
     private static Iterator<Record> mergeTwo(PeekingIterator left, PeekingIterator right) {
-        return new Iterator<>() {
-
-            @Override
-            public boolean hasNext() {
-                return left.hasNext() || right.hasNext();
-            }
-
-            @Override
-            public Record next() {
-                if (!hasNext()) {
-                    throw new NoSuchElementException("No elements");
-                }
-
-                if (!left.hasNext()) {
-                    return right.next();
-                }
-                if (!right.hasNext()) {
-                    return left.next();
-                }
-
-                // checked earlier
-                ByteBuffer leftKey = Objects.requireNonNull(left.peek()).getKey();
-                ByteBuffer rightKey = Objects.requireNonNull(right.peek()).getKey();
-
-                int compareResult = leftKey.compareTo(rightKey);
-                if (compareResult == 0) {
-                    left.next();
-                    return right.next();
-                }
-
-                if (compareResult < 0) {
-                    return left.next();
-                } else {
-                    return right.next();
-                }
-            }
-
-        };
+        return new MergeTwoIterator(left, right);
     }
 
     private static Iterator<Record> filterTombstones(Iterator<Record> iterator) {
@@ -189,7 +149,7 @@ public class LsmDAO implements DAO {
         return new Iterator<>() {
             @Override
             public boolean hasNext() {
-                for (;;) {
+                for (; ; ) {
                     Record peek = delegate.peek();
                     if (peek == null) {
                         return false;
@@ -210,45 +170,5 @@ public class LsmDAO implements DAO {
                 return delegate.next();
             }
         };
-    }
-
-    private static class PeekingIterator implements Iterator<Record> {
-
-        private Record current;
-
-        private final Iterator<Record> delegate;
-
-        public PeekingIterator(Iterator<Record> delegate) {
-            this.delegate = delegate;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return current != null || delegate.hasNext();
-        }
-
-        @Override
-        public Record next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException();
-            }
-            Record now = peek();
-            current = null;
-            return now;
-        }
-
-        public Record peek() {
-            if (current != null) {
-                return current;
-            }
-
-            if (!delegate.hasNext()) {
-                return null;
-            }
-
-            current = delegate.next();
-            return current;
-        }
-
     }
 }
